requirements.txt
python-telegram-bot
git+git://github.com/google/chatbase-python.git
requests
